Changed Navigation Bar Elements
Removed the nav-bar buttons from the add-card view because they were
redundant and implemented undesirable behavior. Now, instead of having
two buttons which return the user to the home screen (with on
presumably saving the data and the other not), there is just a back
button to return to the calling screen and the previously present
add-card button which will be used to save the data.